third_party/blink/web_tests/wpt_internal/html/canvas/drawElement/hit-test/elementsFromPoint.html

<!DOCTYPE html>
<html>
<title>Canvas.setHitTestRegions: hit test regions should appear in elementsFromPoint</title>
<link rel="help" href="https://github.com/WICG/html-in-canvas">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #canvas {
    width: 200px;
    height: 200px;
  }
  div {
    width: 100px;
    height: 100px;
    background: red;
  }
</style>

<canvas id=canvas width="200" height="200" layoutsubtree>
  <div id=a></div>
  <div id=b></div>
  <div id=pointerEventsNone style="pointer-events: none;"></div>
</canvas>

<script>
  test(function () {
    const ctx = canvas.getContext('2d');
    ctx.setHitTestRegions([
      {
        element: a,
        rect: {x: 0, y: 0, width: 100, height: 100}
      },
      {
        element: b,
        rect: {x: 50, y: 50, width: 100, height: 100}
      }
    ]);
    assert_array_equals(document.elementsFromPoint(60, 60),
                        [b, a, canvas, document.body, document.documentElement],
                        'Should have returned a sequence with `[b, a, canvas, body, html]`');
  }, 'Later hit test regions should be hit before earlier ones');

  test(function () {
    const ctx = canvas.getContext('2d');
    ctx.setHitTestRegions([
      {
        element: pointerEventsNone,
        rect: {x: 0, y: 0, width: 100, height: 100}
      },
    ]);
    assert_array_equals(document.elementsFromPoint(60, 60),
                        [canvas, document.body, document.documentElement],
                        'Should have returned a sequence with `[canvas, body, html]`');
  }, 'Pointer-events: none should be excluded from elementsFromPoint');
</script>
</html>
